// src/types/sb_leaderboard_log.ts
export interface LeaderboardSBLog {
  id: string;
  user_id?: bigint;
  rank_score?: number;
  nickname?: string;
  displayname?: string;
  symbol?: string;
  visibility?: string;
  kills?: number;
  deaths?: number;
  score?: bigint;
  damage_dealt?: bigint;
  damage_taken?: bigint;
  matches?: number;
  flight_time?: string;
  wins?: number;
  draws?: number;
  core_captures?: number;
  core_carrier_kills?: number;
  best_lap?: number;
  best_race?: string;
  max_waves?: number;
  num_waves?: number;
  first_blood?: number;
  ace?: number;
  mvp?: number;
  captures?: number;
  shots_fired?: number;
  hits?: number;
  account_media?: string;
  org_media?: string;
  insignia_rank?: number;
  insignia_title?: string;
  rating?: number;
  avg_flight_time?: string;
  score_minute?: number;
  damage_ratio?: number;
  kill_death_ratio?: number;
  losses?: number;
  win_loss_ratio?: number;
  accuracy?: number;
  rank?: number;
  created_at?: bigint;
  map?: string;
}